fix: American amortization base date, take two
Documentation improved. Stopped using the "zero_date" on base date
calculations. The algorithm now relies on the payment date, "due",
instead. This is the correct approach, as specified by the domain
specialists.

Extended the test case for the American amortization system with IPCA.
It now covers the first two payments, instead of only the first one.

Version 4.1.4.